Fix `disemvowel.c` and memory leaks in `disemvowel_test.cpp`
The code now passes Disemvowel.HandleMorrisMinnesota,
Disemvowel.HandlePunctuation, and Disemvowel.HandleNoVowels.
We also fixed the method isVowel to check the letter we send
it from the string. At this time, there is an error of
'invalid write of size 1'.
We also fixed the memory leaks found in `disemvowel_test.cpp`
from tests 'HandleMinnesota', 'HandlePunctuation', and
'HandleLongString'.